fix and note for shader

+++ opengltest/LoadShader.cpp
@@ -49,7 +49,7 @@ static GLuint LoadShaders(const char * vertex_file_path,const char * fragment_fi
     glGetShaderiv(VertexShaderID, GL_COMPILE_STATUS, &Result);
     glGetShaderiv(VertexShaderID, GL_INFO_LOG_LENGTH, &InfoLogLength);
     if ( InfoLogLength > 0 ){
-        std::vector<char> VertexShaderErrorMessage(InfoLogLength+1);
+        std::vector<char> VertexShaderErrorMessage(static_cast<GLuint>(InfoLogLength + 1));
         glGetShaderInfoLog(VertexShaderID, InfoLogLength, NULL, &VertexShaderErrorMessage[0]);
         printf("%s\n", &VertexShaderErrorMessage[0]);
     }
@@ -66,7 +66,7 @@ static GLuint LoadShaders(const char * vertex_file_path,const char * fragment_fi
     glGetShaderiv(FragmentShaderID, GL_COMPILE_STATUS, &Result);
     glGetShaderiv(FragmentShaderID, GL_INFO_LOG_LENGTH, &InfoLogLength);
     if ( InfoLogLength > 0 ){
-        std::vector<char> FragmentShaderErrorMessage(InfoLogLength+1);
+        std::vector<char> FragmentShaderErrorMessage(static_cast<GLuint>(InfoLogLength + 1));
         glGetShaderInfoLog(FragmentShaderID, InfoLogLength, NULL, &FragmentShaderErrorMessage[0]);
         printf("%s\n", &FragmentShaderErrorMessage[0]);
     }
